+++ infrastructure/display-api-service/docker-entrypoint.sh
@@ -2,6 +2,14 @@
 
 set -eux
 
+## Start the PHP FPM process.
+echo "Starting PHP 8.1 FPM"
+
+# first arg is `-f` or `--some-option`
+if [ "${1#-}" != "$1" ]; then
+  set -- php-fpm "$@"
+fi
+
 ## Run templates with configuration.
 /usr/local/bin/confd --onetime --backend env --confdir /etc/confd
 
@@ -10,16 +18,5 @@ composer dump-env prod
 
 ## Warm-up Symfony cache (with the current configuration).
 /var/www/html/bin/console --env=prod cache:warmup
-
-## Set selected composer version. Default version 2.
-if [ ! -z "${COMPOSER_VERSION}" ]; then
-  if [ "${COMPOSER_VERSION}" = "1" ]; then
-    ln -fs /usr/bin/composer1 /home/deploy/bin/composer
-  else
-    ln -fs /usr/bin/composer2 /home/deploy/bin/composer
-  fi
-else
-  ln -fs /usr/bin/composer2 /home/deploy/bin/composer
-fi
 
 exec php-fpm "$@"